DLS: Split Analog Input code
