+++ Watchdogd/Dls/Makefile.am
@@ -3,10 +3,10 @@
 #* Projet WatchDog version 3.0       Gestion d'habitat                      ven 01 aoû 2003 11:37:48 CEST */
 #* Auteur: LEFEVRE Sebastien                                                                              */
 #**********************************************************************************************************/
-SUBDIRS= 
+SUBDIRS=
 
 lib_LTLIBRARIES      = libwatchdog-dls.la
-libwatchdog_dls_la_SOURCES    = plugins.c heure.c Dls_db.c The_dls.c
+libwatchdog_dls_la_SOURCES    = plugins.c heure.c Dls_db.c The_dls.c The_dls_AI.c
 
 libwatchdog_dls_la_LDFLAGS = `pkg-config --libs glib-2.0 gthread-2.0 json-glib-1.0 libsoup-2.4` \
                     `mysql_config --libs` \
